refactor(rules): separate observer failure state
