+++ Assets/Scripts/Rules/Runtime/ObserverFailureState.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: c8063fb83f731714587c768f4f4edbfd
